requirements.txt
# 메인 패키지
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
gunicorn==21.2.0

# 데이터베이스
flask-sqlalchemy==3.1.1
pymysql==1.1.0
cryptography==41.0.7  # PyMySQL 암호화 종속성

# API 클라이언트
anthropic==0.13.0
requests==2.31.0

# 음성 처리
SpeechRecognition==3.10.0
pydub==0.25.1

# 유틸리티
Werkzeug==2.3.7  # Flask와 호환되는 버전
six==1.16.0
python-dateutil==2.8.2

# dotenv 안될수도
python-dotenv==1.0.0

# import anthropic